WIP Big sp<> cleanup
We shouldn't be keeping StateObjects in shared_ptr<>s, that's what StateRef<> is for. And that ends up exposing an implementation detail (that the StateRef<> is effectively backed by an ID lookup and a sp<>), so removing the getSp<> backdoor

TODO:
- Only half done
- Check StateObject::id is set everywhere (Maybe some subclasses define their own id too?) - I think Vehicle::id may be missing?
- Some way of iterating over StateRefMaps as StateRefs<> instead of pair<id, sp<Object>>?

Later:
Normalize object creation - maybe StateRefMap<> should own it, and it can handle the IDs internally? (Either using the next ID integer, or serializer-supplied ID?)
Cleanup weird static StateObject lookup functions?

Investigate if sp<> is really the best store of StateObjects - things hanging onto references after they've been removed from the StateRefMap is already busted in so many ways?